+++ site-cookbooks/network-tools/recipes/default.rb
@@ -1,9 +1,11 @@
 #
-# Cookbook Name:: perl-Net-IP
+# Cookbook Name:: network-tools
 # Recipe:: default
 #
 # Copyright 2015, YOUR_COMPANY_NAME
 #
 # All rights reserved - Do Not Redistribute
 #
-package 'perl-Net-IP'
+node['network-tools']['packages'].each do |pkg|
+	package pkg
+end